+++ caproto/server/common.py
@@ -1,12 +1,26 @@
 from collections import defaultdict, deque, namedtuple, ChainMap, Iterable
 import copy
+import itertools
 import logging
 import sys
 import time
+import weakref
 import caproto as ca
 from caproto import (apply_arr_filter, get_environment_variables,
                      RemoteProtocolError)
 from .._dbr import SubscriptionType
+
+
+# ** Tuning this parameters will affect the servers' performance **
+# ** under high load. **
+# If the queue of subscriptions to has a new update ready within this timeout,
+# we consider ourselves under high load and trade accept some latency for some
+# efficiency.
+HIGH_LOAD_TIMEOUT = 0.01
+# When a batch of subscription updates has this many bytes or more, send it.
+# Decrease this number to improve latency under load; increase it to improve
+# efficiency.
+SUB_BATCH_THRESH = 2**16
 
 
 class DisconnectedCircuit(Exception):
@@ -24,6 +38,40 @@ SubscriptionSpec = namedtuple('SubscriptionSpec', ('db_entry', 'data_type',
 host_endian = ('>' if sys.byteorder == 'big' else '<')
 
 
+class OrderedBoundedSet:
+    # This just implements the part of the set API that we need.
+    def __init__(self, items=None, *, maxlen):
+        if items is None:
+            items = ()
+        self._data = dict.fromkeys(items)
+        self.maxlen = maxlen
+
+    def __len__(self):
+        return len(self._data)
+
+    def add(self, item):
+        self._data[item] = None
+        if len(self) > self.maxlen:
+            self.pop()
+
+    def update(self, items):
+        self._data.update(dict.fromkeys(items))
+        overage = len(self) - self.maxlen
+        for item in itertools.islice(iter(self._data), overage):
+            self._data.pop(item)
+
+    def pop(self):
+        item = next(iter(self._data))
+        self._data.pop(item)
+        return item
+
+    def discard(self, item):
+        try:
+            self._data.pop(item)
+        except KeyError:
+            pass
+
+
 class VirtualCircuit:
     def __init__(self, circuit, client, context):
         self.connected = True
@@ -34,6 +82,16 @@ class VirtualCircuit:
         self.client_hostname = None
         self.client_username = None
         self.subscriptions = defaultdict(deque)
+        self.unexpired_updates = defaultdict(
+            lambda: OrderedBoundedSet(maxlen=ca.MAX_SUBSCRIPTION_BACKLOG))
+        self.most_recent_updates = {}
+        # Subclasses are expected to define:
+        # self.QueueFull = ...
+        # self.command_queue = ...
+        # self.new_command_condition = ...
+        # self.subscription_queue = ...
+        # self.get_from_sub_queue_with_timeout = ...
+        # self.events_on = ...
 
     async def _on_disconnect(self):
         """Executed when disconnection detected"""
@@ -71,7 +129,19 @@ class VirtualCircuit:
 
         commands, _ = self.circuit.recv(bytes_received)
         for c in commands:
-            await self.command_queue.put(c)
+            try:
+                await self.command_queue.put(c)
+            except self.QueueFull:
+                # This client is fast and we are not keeping up. Better to kill
+                # the circuit (and let the client try again) than to let the
+                # whole server be OOM-ed.
+                await self._on_disconnect()
+                raise DisconnectedCircuit()
+                self.log.warning(f"Circuit {self!r} has a large backlog of "
+                                 f"received commands, evidently cannot keep "
+                                 f"with a fast client. Disconnecting circuit "
+                                 f"to avoid letting consume all available "
+                                 f"memory.")
         if not bytes_received:
             await self._on_disconnect()
             raise DisconnectedCircuit()
@@ -163,6 +233,95 @@ class VirtualCircuit:
 
             await self._wake_new_command()
 
+    async def subscription_queue_loop(self):
+        maybe_coroutine = self.events_on.set()
+        # The curio backend makes this an awaitable thing.
+        if maybe_coroutine is not None:
+            await maybe_coroutine
+        commands = deque()
+        commands_bytes = 0
+        num_expired = 0
+        while True:
+            send_now = False
+            try:
+                # We are covering two regimes of operation here. In the "slow
+                # producer" regime, the server is only occasionally sending
+                # updates, and it should optimize for low latency. In the "fast
+                # producer" regime, the server is flooded with updates that it
+                # needs to get out onto the wire as efficiently as possible,
+                # and it should sacrifice some latency in order to batch
+                # requests efficiently.
+                while True:
+                    ref = await self.get_from_sub_queue_with_timeout(HIGH_LOAD_TIMEOUT)
+                    if ref is None:
+                        # We have caught up with the producer. Stop batching,
+                        # and optimize for low latency.
+                        send_now = True
+                        if commands:
+                            # We have accumulated commands while previously in
+                            # the "fast producer" regime. Short-circuit and
+                            # send them.
+                            break
+
+                        # Block here until we have something to send...
+                        ref = await self.subscription_queue.get()
+
+                    command = ref()
+                    if command is None:
+                        # Quota for this subscription has been exceeded.  This
+                        # client is a slow consumer. To avoid letting it get
+                        # behind, drop this message on the floor and move on.
+                        # We are dropping "old news" in favor of prioritizing
+                        # getting the "latest news" out. Note that the
+                        # reference implementation in epics-base, rsrv, does
+                        # the opposite: it drops the new news and sends the old
+                        # news. Jeff Hill has stated clearly that this should
+                        # be considered an implementation detail, not part of
+                        # the specification. The C++ implementation can save
+                        # some memory by discarding the latest updates, but it
+                        # is more useful to discard the oldest updates. Python
+                        # might as well do the more useful thing, given that
+                        # its baseline memory usage is high.
+                        num_expired += 1
+                        continue
+                    self.unexpired_updates[command.subscriptionid].discard(command)
+                    # Accumulate commands into a batch.
+                    commands.append(command)
+                    commands_bytes += len(command)
+                    now = time.monotonic()
+                    if len(commands) == 1:
+                        # Set a dealine by which will must send this oldest
+                        # command in the batch, effecitvely a limit of latency.
+                        deadline = now + HIGH_LOAD_TIMEOUT
+                    elif deadline < now:
+                        send_now = True
+                    # Send the batch if we are in low-latency / slow producer
+                    # mode (send_now=True) or the batch has reached max size.
+                    # But be sure _not_ to send it if it is empty (because all
+                    # the would-be contents were expired.)
+                    if commands and (send_now or commands_bytes > SUB_BATCH_THRESH):
+                        break
+            except self.TaskCancelled:
+                break
+            try:
+                len_commands = len(commands)
+                if num_expired:
+                    self.log.warning("High load. Dropped %d responses.", num_expired)
+                if len_commands > 1:
+                    self.log.info(
+                        "High load. Batched %d commands (%dB) with %.4fs latency.",
+                        len_commands, commands_bytes,
+                        now - deadline + HIGH_LOAD_TIMEOUT)
+                await self.send(*commands)
+            except DisconnectedCircuit:
+                await self._on_disconnect()
+                self.circuit.disconnect()
+                await self.context.circuit_disconnected(self)
+                break
+            commands.clear()
+            commands_bytes = 0
+            num_expired = 0
+
     async def _cull_subscriptions(self, db_entry, func):
         # Iterate through each Subscription, passing each one to func(sub).
         # Collect a list of (SubscriptionSpec, Subscription) for which
@@ -177,6 +336,7 @@ class VirtualCircuit:
                     to_remove.append((sub_spec, sub))
         for sub_spec, sub in to_remove:
             self.subscriptions[sub_spec].remove(sub)
+            self.most_recent_updates.pop(sub.subscriptionid, None)
             self.context.subscriptions[sub_spec].remove(sub)
             self.context.last_dead_band.pop(sub, None)
             self.context.last_sync_edge_update.pop(sub, None)
@@ -338,6 +498,26 @@ class VirtualCircuit:
             return [chan.unsubscribe(command.subscriptionid,
                                      data_type=command.data_type,
                                      data_count=data_count)]
+        elif isinstance(command, ca.EventsOnRequest):
+            # Immediately send most recent updates for all subscriptions.
+            most_recent_updates = list(self.most_recent_updates.values())
+            self.most_recent_updates.clear()
+            if most_recent_updates:
+                await self.send(*most_recent_updates)
+            maybe_coroutine = self.events_on.set()
+            # The curio backend makes this an awaitable thing.
+            if maybe_coroutine is not None:
+                await maybe_coroutine
+        elif isinstance(command, ca.EventsOffRequest):
+            # The client has signaled that it does not think it will be able to
+            # catch up to the backlog. Clear all updates queued to be sent...
+            self.unexpired_updates.clear()
+            # ...and tell the Context that any future updates from ChannelData
+            # should not be added to this circuit's queue until further notice.
+            maybe_coroutine = self.events_on.clear()
+            # The curio backend makes this an awaitable thing.
+            if maybe_coroutine is not None:
+                await maybe_coroutine
         elif isinstance(command, ca.ClearChannelRequest):
             chan, db_entry = get_db_entry()
             await self._cull_subscriptions(
@@ -491,6 +671,7 @@ class Context:
             # This queue receives updates that match the db_entry, data_type
             # and mask ("subscription spec") of one or more subscriptions.
             sub_specs, metadata, values, flags = await self.subscription_queue.get()
+
             subs = []
             for sub_spec in sub_specs:
                 subs.extend(self.subscriptions[sub_spec])
@@ -498,6 +679,7 @@ class Context:
             # We have to make a new response for each channel because each may
             # have a different requested data_count.
             for sub in subs:
+                circuit = sub.circuit
                 s_flags = flags
                 chan = sub.channel
 
@@ -567,10 +749,40 @@ class Context:
                         # Stash this and then send it.
                         self.last_sync_edge_update[sub][sync.s][sync.m] = command
 
-                # Check that the Channel did not close at some point after
-                # this update started its flight.
-                if chan.states[ca.SERVER] is ca.CONNECTED:
-                    await sub.circuit.send(command)
+                # This update will be put at the back of the line of updates to
+                # be sent.
+                #
+                # If len(unexpired_updates[id]) == SUBSCRIPTION_BACKLOG_QUOTA,
+                # then the command at the front of the line will be kicked out.
+                # It is not literally removed from the queue but whenever it
+                # reaches the front of the line it will dropped on the floor
+                # instead of sent. This effectively prioritizes sending the
+                # client "new news" instead of "old news".
+
+                # If this circuit has been sent EventsOff by the client, do not
+                # queue any updates until the client sends EventsOn to signal
+                # that it has caught up. But stash the most recent update for
+                # each subscription, which will immediately send when we turn
+                # events back on.
+                if not circuit.events_on.is_set():
+                    circuit.most_recent_updates[sub.subscriptionid] = command
+                    continue
+
+                # This is an OrderedBoundedSet, a set with a maxlen, containing
+                # only commands for this particular subscription.
+                circuit.unexpired_updates[sub.subscriptionid].add(command)
+
+                # This is a queue with the commands from _all_ subscriptions on
+                # this circuit.
+                try:
+                    await circuit.subscription_queue.put(weakref.ref(command))
+                except self.QueueFull:
+                    # We have hit the overall max for subscription backlog.
+                    circuit.log.warning(
+                        "Critically high EventAddResponse load. Dropping all "
+                        "queued responses on this circuit.")
+                    circuit.subscription_queue.clear()
+                    circuit.unexpired_updates.clear()
 
     async def broadcast_beacon_loop(self):
         self.log.debug('Will send beacons to %r',